updated switch statements for compatibility (BasicClient)
i changed the switch statements from using the '->' operator in the switch statements. to my knowledge, this operator didnt work for me in the past as i was using an older version of java. i updated it so it is more compatible with older versions of java while maintaining logic and usage. feel free to take a look